% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic-core.R
\name{delta_disccovar}
\alias{delta_disccovar}
\title{Perturbation matrix for the discrete-covariate scheme}
\usage{
delta_disccovar(object, p = 1L)
}
\arguments{
\item{object}{A fitted BEZI/BEOI \code{gamlss} model.}

\item{p}{Index of the perturbed covariate in the inflation design matrix.}
}
\value{
A list with the full \code{Delta} and its blocks.
}
\description{
Perturbation matrix for the discrete-covariate scheme
}
