% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Diagnostic residual plots for a BIc model}
\usage{
plot_residuals(object, label_top = 3L)
}
\arguments{
\item{object}{A fitted BEZI/BEOI \code{gamlss} model.}

\item{label_top}{Integer; number of extreme points to label.}
}
\value{
Invisibly returns a list of residual vectors.
}
\description{
Diagnostic residual plots for a BIc model
}
